Another signed comparison issue.
I'd like to rewrite this function to use an owning type for the duplicated
strings, but doing that while keeping the constraints of extern "C" is
tricky. Best would probably be a std::vector<std::vector>> or the equivalent
with std::array, but the lifetimes are dangerous once we start passing
mutable references to C code. For now, we fix the immediate signed
comparison issue and leave this little mess be, comments and all.

+++ soh/soh/OTRGlobals.cpp
@@ -347,7 +347,7 @@ OTRGlobals::OTRGlobals() {
     // This is in OTRGlobals right now because this is a place that will only ever be run once at the beginning of startup.
     // We should probably find some code in db_camera that does initialization and only run once, and then dealloc on deinitialization.
     cameraStrings = (char**)malloc(sizeof(constCameraStrings));
-    for (int32_t i = 0; i < sizeof(constCameraStrings) / sizeof(char*); i++) {
+    for (size_t i = 0; i < sizeof(constCameraStrings) / sizeof(char*); i++) {
         // OTRTODO: never deallocated...
         auto dup = strdup(constCameraStrings[i]);
         cameraStrings[i] = dup;
